% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_plot}
\alias{test_plot}
\title{Mimic Base R break}
\usage{
test_plot(
  df,
  group_var,
  value_var,
  fill_var,
  p_label = "p.signif",
  detailed = F,
  hide_ns = T
)
}
\arguments{
\item{x}{string with which to prefix names of \code{ggplot2} functions in order to name the pipe-enabled functions. default: "add_".
Note: You could set this the empty string, in which case the new functions would mask the name of the library function}

\item{y}{func_regex Regular expression to filter the list of ggplot functions to make pipe-enabled.  The default regex will capture all}
}
\description{
Mimic Base R break
}
